+++ src/pydbm/database/__init__.py
@@ -1,3 +1,5 @@
 from pydbm.database.manager import DatabaseManager
 
-__all__ = ("DatabaseManager",)
+__all__ = (
+    "DatabaseManager",
+)