+++ src/pydbm/models/validators/__init__.py
@@ -1,16 +1,14 @@
+import datetime
+
 from pydbm.models.validators.builtin_types import (
     validate_bool,
     validate_bytes,
     validate_date,
     validate_datetime,
-    validate_dict,
     validate_float,
     validate_int,
-    validate_list,
-    validate_nonetype,
-    validate_set,
+    validate_none,
     validate_str,
-    validate_tuple,
 )
 from pydbm.models.validators.compare import validate_max_value, validate_min_value
 
@@ -19,14 +17,23 @@ __all__ = (
     "validate_bytes",
     "validate_date",
     "validate_datetime",
-    "validate_dict",
     "validate_float",
     "validate_int",
-    "validate_list",
     "validate_max_value",
     "validate_min_value",
-    "validate_nonetype",
-    "validate_set",
+    "validate_none",
     "validate_str",
-    "validate_tuple",
+    "validator_mapping",
 )
+
+
+validator_mapping = {
+    bool: validate_bool,
+    bytes: validate_bytes,
+    datetime.date: validate_date,
+    datetime.datetime: validate_datetime,
+    float: validate_float,
+    int: validate_int,
+    None: validate_none,
+    str: validate_str,
+}