+++ src/pydbm/models/validators/builtin_types.py
@@ -6,14 +6,10 @@ __all__ = (
     "validate_bytes",
     "validate_date",
     "validate_datetime",
-    "validate_dict",
     "validate_float",
     "validate_int",
-    "validate_list",
-    "validate_nonetype",
-    "validate_set",
+    "validate_none",
     "validate_str",
-    "validate_tuple",
 )
 
 
@@ -37,11 +33,6 @@ def validate_datetime(value: typing.Any) -> None:
         raise ValueError("It must be datetime")
 
 
-def validate_dict(value: typing.Any) -> None:
-    if value.__class__ is not dict:
-        raise ValueError("It must be dict")
-
-
 def validate_float(value: typing.Any) -> None:
     if value.__class__ is not float:
         raise ValueError("It must be float")
@@ -52,26 +43,11 @@ def validate_int(value: typing.Any) -> None:
         raise ValueError("It must be int")
 
 
-def validate_list(value: typing.Any) -> None:
-    if value.__class__ is not list:
-        raise ValueError("It must be list")
-
-
-def validate_nonetype(value: typing.Any) -> None:
+def validate_none(value: typing.Any) -> None:
     if value is not None:
         raise ValueError("It must be None")
-
-
-def validate_set(value: typing.Any) -> None:
-    if value.__class__ is not set:
-        raise ValueError("It must be set")
 
 
 def validate_str(value: typing.Any) -> None:
     if value.__class__ is not str:
         raise ValueError("It must be str")
-
-
-def validate_tuple(value: typing.Any) -> None:
-    if value.__class__ is not tuple:
-        raise ValueError("It must be tuple")